Move non-editor-related components to atoms.
