+++ packages/editor/src/client/blocks/richTextBlock/richTextBlock.tsx
@@ -5,9 +5,9 @@ import {withHistory} from 'slate-history'
 import {withReact, ReactEditor, Editable, Slate} from 'slate-react'
 import {BlockProps} from '../../atoms/blockList'
 import {EmojiPicker} from '../../atoms/emojiPicker'
-import {Toolbar, ToolbarDivider} from '../../atoms/toolbar'
+import {Toolbar, ToolbarDivider, H1Icon, H2Icon, H3Icon} from '../../atoms/toolbar'
 import {RichTextBlockValue} from '../types'
-import {FormatButton, H1Icon, H2Icon, H3Icon, FormatIconButton, SlateSubMenuButton} from './buttons'
+import {FormatButton, FormatIconButton, SlateSubMenuButton} from './buttons'
 import {renderElement, renderLeaf, withRichText} from './editor'
 import {BlockFormat, TextFormat} from './formats'
 import {LinkFormatButton, RemoveLinkFormatButton} from './linkButton'
